// WADNR.Web/src/app/shared/generated/model/bulk-tag-projects-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class BulkTagProjectsRequest { 
    TagName?: string | null;
    ProjectIDs?: Array<number> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface BulkTagProjectsRequestForm { 
    TagName?: FormControl<string>;
    ProjectIDs?: FormControl<Array<number>>;
}

export class BulkTagProjectsRequestFormControls { 
    public static TagName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}